requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.1
bracex==2.6
branca==0.8.2
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
deepagents==0.3.4
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.2
filetype==1.2.0
folium==0.20.0
fsspec==2025.12.0
google-auth==2.47.0
google-genai==1.57.0
gradio==6.2.0
gradio_client==2.0.2
groovy==0.1.2
groq==0.37.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.4
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.3
langchain-anthropic==1.3.1
langchain-core==1.2.6
langchain-google-genai==4.1.3
langchain-groq==1.1.1
langchain-openai==1.1.7
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.2
logging==0.4.9.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.0
openai==2.14.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
safehttpx==0.1.7
scipy==1.16.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
statsmodels==0.14.6
tenacity==9.1.2
tiktoken==0.12.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
wcmatch==10.1
websockets==15.0.1
xxhash==3.6.0
xyzservices==2025.11.0
zstandard==0.25.0
